Fix compiler warnings in our own code
Deprecated QVariant::type(), signed/unsigned comparisons against
taglib's size(), unused locals, an unchecked nodiscard return and the
deprecated Q_OS_MACX. AlbumCoverManagerList::mimeData took the list by
value, which stopped overriding QListWidget's since Qt6 changed it to a
const reference, so dragging albums out of the cover manager lost the
songs.

+++ ext/libclementine-tagreader/cloudstream.cpp
@@ -56,13 +56,13 @@ bool CloudStream::CheckCache(int start, int end) {
 }
 
 void CloudStream::FillCache(int start, TagLib::ByteVector data) {
-  for (int i = 0; i < data.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(data.size()); ++i) {
     cache_.set(start + i, data[i]);
   }
 }
 
 TagLib::ByteVector CloudStream::GetCached(int start, int end) {
-  const uint size = end - start + 1;
+  const int size = end - start + 1;
   TagLib::ByteVector ret(size);
   for (int i = 0; i < size; ++i) {
     ret[i] = cache_.get(start + i);
